Move calendar components into "calendar" folder
