+++ app/calendar/caldav/index.ts
@@ -2,7 +2,7 @@ import * as urllib from "urllib"
 import * as ical from "node-ical"
 import _ from "lodash"
 import { v4 as uuidv4 } from "uuid"
-import type { CalendarService, CreateEventBooking } from "app/calendar-service"
+import type { CalendarService, CreateEventBooking } from "app/calendar/calendar-service"
 import { ConnectedCalendar } from "db"
 import passwordEncryptor from "app/users/password-encryptor"
 import { addMinutes } from "date-fns"
