+++ app/calendar/calendar-service.ts
@@ -1,6 +1,6 @@
 import { Booking, ConnectedCalendar, Meeting, User } from "db"
 import { ExternalEvent, getCalendarService as getCalDavCalendarService } from "./caldav"
-import { getCalendarService as getGoogleCalendarService } from "app/googlecalendar/googlecalendar"
+import { getCalendarService as getGoogleCalendarService } from "app/calendar/googlecalendar/googlecalendar"
 
 export type CreateEventBooking = Pick<Booking, "startDateUTC" | "inviteeEmail"> & {
   meeting: Pick<Meeting, "duration" | "location" | "name" | "description"> & {
